add site-load.js
Перенес функции отрисовки по кнопкам в хэдере в отдельный файл js
