+++ src/index.js
@@ -63,6 +63,7 @@ import { GENRES_STORAGE } from './js/fetchMovies';
 import movieListMarkupHbs from './templates/movie-list.hbs';
 import modalMarkupHbs from './templates/modal.hbs';
 import getRefs from './js/get-refs';
+import { onHomePageLoading, onLibraryPageLoading } from './js/site-load';
 
 var debounce = require('lodash.debounce');
 const DEBOUNCE_DELAY = 600;
@@ -73,14 +74,12 @@ let PAGE = 1;
 let totalPages = undefined;
 const refs = getRefs();
 
-
 // ===================Ищет популярные=====================
 
 refs.sitePage.classList.add('js-page-header__home');
 refs.homePageBtn.classList.add('js-navigation__button--current');
 refs.libPageBtnNav.classList.add('js-visually-hidden');
 refs.libBtnWatched.classList.add('js-library__button--current');
-
 
 refs.searchBox.addEventListener('input', debounce(onSearchInputs, DEBOUNCE_DELAY));
 refs.siteLogo.addEventListener('click', onHomePageLoading);
@@ -102,7 +101,6 @@ fetchMarkupPopularityForWeek();
 // }
 
 function fetchMarkupPopularityForWeek() {
-
   PAGE = 1;
   fetchPopularity(PAGE)
     .then(processGenres)
@@ -113,7 +111,6 @@ function fetchMarkupPopularityForWeek() {
     });
 }
 
-
 // ==============================Открывает-Закрывает Модалку==========================
 refs.movieList.addEventListener('click', onClickInItem);
 refs.movieModal.addEventListener('click', onClickBackdrop);
@@ -122,7 +119,6 @@ function onClickInItem(e) {
   if (e.target.className === 'movie-img') {
     onOpenModal();
   }
-
 }
 
 function onOpenModal() {
@@ -130,7 +126,6 @@ function onOpenModal() {
   refs.movieModal.classList.remove('is-hidden');
   modalMarkup();
 }
-
 
 function onCloseModal() {
   window.removeEventListener('keydown', onEscKeyDown);
@@ -264,22 +259,3 @@ const observer = new IntersectionObserver(onEntry, {
   rootMargin: '150px',
 });
 observer.observe(refs.sentinel);
-
-
-//=====подключение стилей при навигации по страницам=====
-function onHomePageLoading() {
-  refs.sitePage.classList.replace('js-page-header__library', 'js-page-header__home');
-  refs.libPageBtn.classList.remove('js-navigation__button--current');
-  refs.homePageBtn.classList.add('js-navigation__button--current');
-  refs.homePageForm.classList.remove('js-visually-hidden');
-  refs.libPageBtnNav.classList.add('js-visually-hidden');
-}
-
-function onLibraryPageLoading() {
-  refs.sitePage.classList.replace('js-page-header__home', 'js-page-header__library');
-  refs.homePageBtn.classList.remove('js-navigation__button--current');
-  refs.libPageBtn.classList.add('js-navigation__button--current');
-  refs.libPageBtnNav.classList.remove('js-visually-hidden');
-  refs.homePageForm.classList.add('js-visually-hidden');
-}
-
