+++ src/templates/movie-list.hbs
@@ -78,7 +78,7 @@
         <p style="margin-top: 0;">
         <span class="movie-item__year">{{release_date.[0]}}{{release_date.[1]}}{{release_date.[2]}}{{release_date.[3]}}</span>
         {{!-- можно скрыть --}}
-        <span class="movie-item__vote-average">{{vote_average}}</span>
+        <span class="movie-item__vote-average js-visually-hidden">{{vote_average}}</span>
         </p>
         </div>
 </div>
